Experiment with /blog/something. Experiment Successful for GetMappings.
Now, any blog related page, like the post details page will be shown as "../blog/post.." instead of "../post..". This reduces the number of config pieces required in the security config file. It also means that login and logout actions in forms need to be "../login" and "../logout" to get to the main mapping. Icons have been added to the main page for navigation to the respective dashboards and to logout.